% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyigos-package.R
\docType{package}
\name{manyigos-package}
\alias{manyigos}
\alias{manyigos-package}
\title{manyigos: Intergovernmental Organisations for manydata}
\description{
manyigos provides data on intergovernmental organisations in the international system across time. The package combines and cross-references several existing datasets consistent with the manydata package aims.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/globalgov/manyigos}
  \item Report bugs at \url{https://github.com/globalgov/manyigos/issues}
}

}
\author{
\strong{Maintainer}: James Hollway \email{james.hollway@graduateinstitute.ch} (\href{https://orcid.org/0000-0002-8361-9647}{ORCID}) (IHEID) [contributor]

Other contributors:
\itemize{
  \item Jael Tan (\href{https://orcid.org/0000-0002-6234-9764}{ORCID}) (IHEID) [contributor]
  \item Henrique Sposito (\href{https://orcid.org/0000-0003-3420-6085}{ORCID}) (IHEID) [contributor]
}

}
\keyword{internal}
