% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyigos-organizations.R
\docType{data}
\name{organizations}
\alias{organizations}
\title{organizations database documentation}
\format{
The organizations database is a list that contains the
following 7 datasets: COW, VIGO, DIGO, IIGO, TRANSACCESS, YIO, MIA.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{COW: }{A dataset with 16139 observations and the following
18 variables: igoID, Title, cowigoNR, Beg, End, Year, orgname, dead,
integrated, replaced, igocode, version, accuracyofpre1965membershipdates,
sourcesandnotes, imputed, political, social, economic.}
\item{VIGO: }{A dataset with 1662 observations and the following
26 variables: igoID, Title, Beg, Year, vitality, zombie, life, dead,
numb_mids_sum, mccall, pr_integration, aut3, s3un, Europe, LAmerica, MENA,
Asia, NAmCarib, Africa, compdummy, age, competition, hardrel, itrans,
desec5f, lngdpsum1.}
\item{DIGO: }{A dataset with 561 observations and the following
29 variables: igoID, Title, Beg, End, Death, region, memb_form,
members_censor, Mandate/Function, Living, Scope, function_c, scope_c,
region_c, lifespan, event, event_1, life_1, WWI, WWII, WW,
full_membership_start, asso_membership_start, observer_start,
missing_data_start, full_membership_end, asso_membership_end, observer_end,
missing_data_end.}
\item{IIGO: }{A dataset with 149 observations and the following
12 variables: igoID, Label, Beg, End, End_reason, Regional, Nbr., Domain,
Nbr State Mbrs 2017 (or end), Ind. Sec. Year, Formal Treaty Year,
Follow-on IO.}
\item{TRANSACCESS: }{A dataset with 9007 observations and the following
32 variables: igoID, Title, igobodyID, igoBody, Beg, Year, access, ID, ID_IO,
ID_Body, openbody, multiple_arr, range_1, depth_1, permanence_1,
codification_1, range_2, depth_2, permanence_2, codification_2, range_3,
depth_3, permanence_3, codification_3, range_4, depth_4, permanence_4,
codification_4, range_5, depth_5, permanence_5, codification_5.}
\item{YIO: }{A dataset with 75115 observations and the following
4 variables: Title, Abbreviation, Beg, Country.}
\item{MIA: }{A dataset with 3961 observations and the following
54 variables: igoID, Title, cowigoNR, Beg, Year, initial, end, typeI,
contract, DIMENSIONS, delegation, pooling, DS_sum_st, COMPONENTS, delaccess,
delsuspens, delconstit, delbudget, delcompliance, delpolicy, del_agenda,
del_final, poolaccess, poolsuspens, poolconstit, poolbudget, poolcompliance,
poolpolicy, poolagenda, poolfinal, coverage1, thirdparty1, tribunal1,
binding1, nonstate1, remedy1, preliminary1, coverage2, thirdparty2,
tribunal2, binding2, nonstate2, remedy2, preliminary2, DS_sum1, DS_sum2,
DS_sum3, coverage3, thirdparty3, tribunal3, binding3, nonstate3, remedy3,
preliminary3.}
}
}
\usage{
organizations
}
\description{
organizations database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $COW
#> ---------------------------------------------------------------------------------------------
#> |          Column Name           |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ---------------------------------------------------------------------------------------------
#> |             igoID              |  character  |           16139|     0     |       0       |
#> |             Title              |  character  |           16139|     0     |       0       |
#> |            cowigoNR            |  character  |           16139|     0     |       0       |
#> |              Beg               |    mdate    |           16139|     0     |       0       |
#> |              End               |    mdate    |           16139|   12173   |     75.43     |
#> |              Year              |    mdate    |           16139|     0     |       0       |
#> |            orgname             |  character  |           16139|     0     |       0       |
#> |              dead              |  character  |           16139|     0     |       0       |
#> |           integrated           |  character  |           16139|     0     |       0       |
#> |            replaced            |  character  |           16139|     0     |       0       |
#> |            igocode             |  character  |           16139|   6967    |     43.17     |
#> |            version             |  character  |           16139|     0     |       0       |
#> |accuracyofpre1965membershipdates|  character  |           16139|    159    |     0.99      |
#> |        sourcesandnotes         |  character  |           16139|   15981   |     99.02     |
#> |            imputed             |  character  |           16139|     0     |       0       |
#> |           political            |  character  |           16139|     0     |       0       |
#> |             social             |  character  |           16139|     0     |       0       |
#> |            economic            |  character  |           16139|     0     |       0       |
#> ---------------------------------------------------------------------------------------------
#> 
#> 
#> $VIGO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     igoID     |  character  |            1662|     0     |       0       |
#> |     Title     |  character  |            1662|     0     |       0       |
#> |      Beg      |    mdate    |            1662|     0     |       0       |
#> |     Year      |    mdate    |            1662|     0     |       0       |
#> |   vitality    |  character  |            1662|     0     |       0       |
#> |    zombie     |  character  |            1662|     0     |       0       |
#> |     life      |  character  |            1662|     0     |       0       |
#> |     dead      |  character  |            1662|     0     |       0       |
#> | numb_mids_sum |  character  |            1662|    231    |     13.9      |
#> |    mccall     |  character  |            1662|   1049    |     63.12     |
#> |pr_integration |  character  |            1662|     9     |     0.54      |
#> |     aut3      |  character  |            1662|    269    |     16.19     |
#> |     s3un      |  character  |            1662|    30     |     1.81      |
#> |    Europe     |  character  |            1662|     9     |     0.54      |
#> |   LAmerica    |  character  |            1662|     9     |     0.54      |
#> |     MENA      |  character  |            1662|     9     |     0.54      |
#> |     Asia      |  character  |            1662|     9     |     0.54      |
#> |   NAmCarib    |  character  |            1662|     9     |     0.54      |
#> |    Africa     |  character  |            1662|     9     |     0.54      |
#> |   compdummy   |  character  |            1662|     9     |     0.54      |
#> |      age      |  character  |            1662|     0     |       0       |
#> |  competition  |  character  |            1662|     0     |       0       |
#> |    hardrel    |  character  |            1662|    589    |     35.44     |
#> |    itrans     |  character  |            1662|     0     |       0       |
#> |    desec5f    |  character  |            1662|     0     |       0       |
#> |   lngdpsum1   |  character  |            1662|    508    |     30.57     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $DIGO
#> ----------------------------------------------------------------------------------
#> |     Column Name     |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------------
#> |        igoID        |  character  |             561|     0     |       0       |
#> |        Title        |  character  |             561|     0     |       0       |
#> |         Beg         |    mdate    |             561|     0     |       0       |
#> |         End         |    mdate    |             561|     0     |       0       |
#> |        Death        |  character  |             561|    349    |     62.21     |
#> |       region        |  character  |             561|     1     |     0.18      |
#> |      memb_form      |  character  |             561|     1     |     0.18      |
#> |   members_censor    |  character  |             561|    12     |     2.14      |
#> |  Mandate/Function   |  character  |             561|     2     |     0.36      |
#> |       Living        |  character  |             561|     1     |     0.18      |
#> |        Scope        |  character  |             561|     2     |     0.36      |
#> |     function_c      |  character  |             561|     2     |     0.36      |
#> |       scope_c       |  character  |             561|     2     |     0.36      |
#> |      region_c       |  character  |             561|     1     |     0.18      |
#> |      lifespan       |  character  |             561|     1     |     0.18      |
#> |        event        |  character  |             561|     3     |     0.53      |
#> |       event_1       |  character  |             561|     4     |     0.71      |
#> |       life_1        |  character  |             561|     3     |     0.53      |
#> |         WWI         |  character  |             561|     2     |     0.36      |
#> |        WWII         |  character  |             561|     1     |     0.18      |
#> |         WW          |  character  |             561|     1     |     0.18      |
#> |full_membership_start|  character  |             561|    18     |     3.21      |
#> |asso_membership_start|  character  |             561|    17     |     3.03      |
#> |   observer_start    |  character  |             561|    17     |     3.03      |
#> | missing_data_start  |  character  |             561|    17     |     3.03      |
#> | full_membership_end |  character  |             561|     1     |     0.18      |
#> | asso_membership_end |  character  |             561|     1     |     0.18      |
#> |    observer_end     |  character  |             561|     1     |     0.18      |
#> |  missing_data_end   |  character  |             561|     1     |     0.18      |
#> ----------------------------------------------------------------------------------
#> 
#> 
#> $IIGO
#> -----------------------------------------------------------------------------------------
#> |        Column Name         |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -----------------------------------------------------------------------------------------
#> |           igoID            |  character  |             149|     2     |     1.34      |
#> |           Label            |  character  |             149|     0     |       0       |
#> |            Beg             |    mdate    |             149|     0     |       0       |
#> |            End             |    mdate    |             149|    98     |     65.77     |
#> |         End_reason         |  character  |             149|     0     |       0       |
#> |          Regional          |  character  |             149|     0     |       0       |
#> |            Nbr.            |  character  |             149|     0     |       0       |
#> |           Domain           |  character  |             149|     3     |     2.01      |
#> |Nbr State Mbrs 2017 (or end)|  character  |             149|     0     |       0       |
#> |       Ind. Sec. Year       |  character  |             149|    117    |     78.52     |
#> |     Formal Treaty Year     |  character  |             149|    134    |     89.93     |
#> |        Follow-on IO        |  character  |             149|     1     |     0.67      |
#> -----------------------------------------------------------------------------------------
#> 
#> 
#> $TRANSACCESS
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     igoID     |  character  |            9007|     0     |       0       |
#> |     Title     |  character  |            9007|     0     |       0       |
#> |   igobodyID   |  character  |            9007|     0     |       0       |
#> |    igoBody    |  character  |            9007|     0     |       0       |
#> |      Beg      |    mdate    |            9007|     0     |       0       |
#> |     Year      |    mdate    |            9007|     0     |       0       |
#> |    access     |  character  |            9007|     0     |       0       |
#> |      ID       |  character  |            9007|     0     |       0       |
#> |     ID_IO     |  character  |            9007|     0     |       0       |
#> |    ID_Body    |  character  |            9007|     0     |       0       |
#> |   openbody    |  character  |            9007|     0     |       0       |
#> | multiple_arr  |  character  |            9007|     0     |       0       |
#> |    range_1    |  character  |            9007|     0     |       0       |
#> |    depth_1    |  character  |            9007|     0     |       0       |
#> | permanence_1  |  character  |            9007|     0     |       0       |
#> |codification_1 |  character  |            9007|     0     |       0       |
#> |    range_2    |  character  |            9007|     0     |       0       |
#> |    depth_2    |  character  |            9007|     0     |       0       |
#> | permanence_2  |  character  |            9007|     0     |       0       |
#> |codification_2 |  character  |            9007|     0     |       0       |
#> |    range_3    |  character  |            9007|     0     |       0       |
#> |    depth_3    |  character  |            9007|     0     |       0       |
#> | permanence_3  |  character  |            9007|     0     |       0       |
#> |codification_3 |  character  |            9007|     0     |       0       |
#> |    range_4    |  character  |            9007|     0     |       0       |
#> |    depth_4    |  character  |            9007|     0     |       0       |
#> | permanence_4  |  character  |            9007|     0     |       0       |
#> |codification_4 |  character  |            9007|     0     |       0       |
#> |    range_5    |  character  |            9007|     0     |       0       |
#> |    depth_5    |  character  |            9007|     0     |       0       |
#> | permanence_5  |  character  |            9007|     0     |       0       |
#> |codification_5 |  character  |            9007|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $YIO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     Title     |  character  |           75115|     0     |       0       |
#> | Abbreviation  |  character  |           75115|   29296   |      39       |
#> |      Beg      |    mdate    |           75115|   23141   |     30.81     |
#> |    Country    |  character  |           75115|   28101   |     37.41     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $MIA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     igoID     |  character  |            3961|     0     |       0       |
#> |     Title     |  character  |            3961|     0     |       0       |
#> |   cowigoNR    |  character  |            3961|     0     |       0       |
#> |      Beg      |    mdate    |            3961|     0     |       0       |
#> |     Year      |    mdate    |            3961|     0     |       0       |
#> |    initial    |  character  |            3961|   3884    |     98.06     |
#> |      end      |  character  |            3961|   3885    |     98.08     |
#> |     typeI     |  character  |            3961|     0     |       0       |
#> |   contract    |  character  |            3961|     0     |       0       |
#> |  DIMENSIONS   |  character  |            3961|   3961    |      100      |
#> |  delegation   |  character  |            3961|     0     |       0       |
#> |    pooling    |  character  |            3961|     0     |       0       |
#> |   DS_sum_st   |  character  |            3961|     0     |       0       |
#> |  COMPONENTS   |  character  |            3961|   3961    |      100      |
#> |   delaccess   |  character  |            3961|     0     |       0       |
#> |  delsuspens   |  character  |            3961|     0     |       0       |
#> |  delconstit   |  character  |            3961|     0     |       0       |
#> |   delbudget   |  character  |            3961|     0     |       0       |
#> | delcompliance |  character  |            3961|     0     |       0       |
#> |   delpolicy   |  character  |            3961|     0     |       0       |
#> |  del_agenda   |  character  |            3961|     0     |       0       |
#> |   del_final   |  character  |            3961|     0     |       0       |
#> |  poolaccess   |  character  |            3961|     0     |       0       |
#> |  poolsuspens  |  character  |            3961|     0     |       0       |
#> |  poolconstit  |  character  |            3961|     0     |       0       |
#> |  poolbudget   |  character  |            3961|     0     |       0       |
#> |poolcompliance |  character  |            3961|     0     |       0       |
#> |  poolpolicy   |  character  |            3961|     0     |       0       |
#> |  poolagenda   |  character  |            3961|     0     |       0       |
#> |   poolfinal   |  character  |            3961|     0     |       0       |
#> |   coverage1   |  character  |            3961|     8     |      0.2      |
#> |  thirdparty1  |  character  |            3961|     8     |      0.2      |
#> |   tribunal1   |  character  |            3961|     8     |      0.2      |
#> |   binding1    |  character  |            3961|     8     |      0.2      |
#> |   nonstate1   |  character  |            3961|     8     |      0.2      |
#> |    remedy1    |  character  |            3961|     8     |      0.2      |
#> | preliminary1  |  character  |            3961|     8     |      0.2      |
#> |   coverage2   |  character  |            3961|    31     |     0.78      |
#> |  thirdparty2  |  character  |            3961|    31     |     0.78      |
#> |   tribunal2   |  character  |            3961|    31     |     0.78      |
#> |   binding2    |  character  |            3961|    31     |     0.78      |
#> |   nonstate2   |  character  |            3961|    31     |     0.78      |
#> |    remedy2    |  character  |            3961|    31     |     0.78      |
#> | preliminary2  |  character  |            3961|    31     |     0.78      |
#> |    DS_sum1    |  character  |            3961|     0     |       0       |
#> |    DS_sum2    |  character  |            3961|     0     |       0       |
#> |    DS_sum3    |  character  |            3961|     0     |       0       |
#> |   coverage3   |  character  |            3961|   3958    |     99.92     |
#> |  thirdparty3  |  character  |            3961|   3958    |     99.92     |
#> |   tribunal3   |  character  |            3961|   3958    |     99.92     |
#> |   binding3    |  character  |            3961|   3958    |     99.92     |
#> |   nonstate3   |  character  |            3961|   3958    |     99.92     |
#> |    remedy3    |  character  |            3961|   3958    |     99.92     |
#> | preliminary3  |  character  |            3961|   3958    |     99.92     |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
